transformed data to float32

+++ ndvi.py
@@ -9,13 +9,13 @@ def make_ndvi(before,after):
     red_after = after[:,:,0]
     nir_after = after[:,:,1]
 
-    ndvi_before = np.empty(red_before.shape,dtype = np.half)
+    ndvi_before = np.empty(red_before.shape,dtype = np.float32)
     check_before = np.logical_or(red_before>0,nir_before>0)
-    ndvi_before = np.where(check_before, (nir_before-red_before)/(nir_before+red_before),0).astype(np.half)
+    ndvi_before = np.where(check_before, (nir_before-red_before)/(nir_before+red_before),0).astype(np.float32)
     
-    ndvi_after = np.empty(red_after.shape, dtype = np.uint16)
+    ndvi_after = np.empty(red_after.shape, dtype = np.float32)
     check_after = np.logical_or(red_after>0, nir_after>0)
-    ndvi_after = np.where(check_after, (nir_after-red_after)/(nir_after+red_after),0).astype(np.half)
+    ndvi_after = np.where(check_after, (nir_after-red_after)/(nir_after+red_after),0).astype(np.float32)
     
     return ndvi_before,ndvi_after
